requirements.txt
aiosqlite~=0.22.0
fastapi~=0.124.4
uvicorn~=0.38.0
sqlalchemy~=2.0.45
pydantic~=2.12.5

pytest~=9.0.2
httpx~=0.28.1
factory-boy~=3.3.3
black~=25.12.0
Faker~=38.2.0
isort~=7.0.0
flake8~=7.3.0
flake8-bugbear~=25.11.29
flake8-pie~=0.16.0
mypy~=1.19.1
anyio~=4.12.0
pytest-asyncio~=1.3.0
pytest-tornasync~=0.6.0.post2
